requirements.txt
amqp==5.3.1
attrs==25.3.0
billiard==4.2.1
celery==5.4.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
crontab==1.0.1
cryptography==44.0.2
greenlet==3.1.1
h11==0.14.0
idna==3.10
kombu==5.5.2
MouseInfo==0.1.3
numpy==2.2.4
outcome==1.3.0.post0
packaging==24.2
pandas==2.2.3
pdfminer.six==20250327
pdfplumber==0.11.6
pillow==11.1.0
playwright==1.51.0
prompt_toolkit==3.0.50
PyAutoGUI==0.9.54
pycparser==2.22
pyee==12.1.1
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pypdfium2==4.30.1
pyperclip==1.9.0
PyRect==0.2.0
PyScreeze==1.0.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python3-xlib==0.15
pytweening==1.2.0
pytz==2025.2
redis==5.2.1
requests==2.32.3
selenium==4.30.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
trio==0.29.0
trio-websocket==0.12.2
typing_extensions==4.13.0
tzdata==2025.2
urllib3==2.3.0
vine==5.1.0
wcwidth==0.2.13
webdriver-manager==4.0.2
websocket-client==1.8.0
wsproto==1.2.0
